refactor: Split the RealPlugin class into several files
The structure is nearly unchanged (partial class)
