+++ Source/Triggernometry/TriggernometryPlugin.csproj
@@ -395,6 +395,17 @@
     <Compile Include="ObsController.cs" />
     <Compile Include="LiveSplitController.cs" />
     <Compile Include="Parser.cs" />
+    <Compile Include="RealPlugin\RealPlugin.Helper.cs" />
+    <Compile Include="RealPlugin\RealPlugin.Config.cs" />
+    <Compile Include="RealPlugin\RealPlugin.Scripting.cs" />
+    <Compile Include="RealPlugin\RealPlugin.I18n.cs" />
+    <Compile Include="RealPlugin\RealPlugin.Trigger.cs" />
+    <Compile Include="RealPlugin\RealPlugin.InternalLog.cs" />
+    <Compile Include="RealPlugin\RealPlugin.ActionQueue.cs" />
+    <Compile Include="RealPlugin\RealPlugin.Aura.cs" />
+    <Compile Include="RealPlugin\RealPlugin.Audio.cs" />
+    <Compile Include="RealPlugin\RealPlugin.Update.cs" />
+    <Compile Include="RealPlugin\RealPlugin.NamedCallback.cs" />
     <Compile Include="Scarborough\OverrideHelper.cs" />
     <Compile Include="PluginBridges\BridgeFFXIV.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
@@ -402,7 +413,7 @@
       <DesignTime>True</DesignTime>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
-    <Compile Include="RealPlugin.cs" />
+    <Compile Include="RealPlugin\RealPlugin.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Repository.cs" />
     <Compile Include="Forms\RepositoryForm.cs">
